track dependency on complete type of an object definition
git-svn-id: https://cpparch.googlecode.com/svn/trunk@483 174230dc-fbae-11de-8037-3961252a365c

+++ cppparse/test/include/incomplete_sizeoftype.h
@@ -4,7 +4,7 @@
 struct S;
 namespace SizeOfType
 {
-	size_t size = sizeof(S);
+	int size = sizeof(S);
 }
 
 #endif
